ai_models/requirements-training.txt
# Training Requirements for Bird Image/GIF Generation
# GPU-enabled environment required (CUDA 11.8+)

# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
--extra-index-url https://download.pytorch.org/whl/cu118

# Diffusion Models
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0  # LoRA training
bitsandbytes>=0.41.0  # Quantization

# Image Generation
safetensors>=0.4.0
omegaconf>=2.3.0
invisible-watermark>=0.2.0
xformers>=0.0.23  # Memory-efficient attention

# ControlNet & Extensions
controlnet-aux>=0.0.7
mediapipe>=0.10.0  # For pose detection

# Video/GIF Generation
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
av>=11.0.0
decord>=0.6.0

# Data Processing
Pillow>=10.1.0
opencv-python>=4.8.0
albumentations>=1.3.1
numpy>=1.24.0
scipy>=1.11.0

# Dataset Management
datasets>=2.15.0
huggingface-hub>=0.19.0
boto3>=1.34.0  # S3 access
smart-open>=6.4.0

# Experiment Tracking & MLOps
wandb>=0.16.0
tensorboard>=2.15.0
mlflow>=2.9.0

# Prompt Engineering
openai>=1.6.0  # For GPT-4 prompt generation
anthropic>=0.8.0  # Claude alternative
tiktoken>=0.5.0

# Optimization
onnx>=1.15.0
onnxruntime-gpu>=1.16.0
tensorrt>=8.6.0  # NVIDIA optimization

# Training Utilities
einops>=0.7.0
timm>=0.9.0  # Vision models
pytorch-lightning>=2.1.0
deepspeed>=0.12.0  # Distributed training

# Quality Metrics
lpips>=0.1.4  # Perceptual similarity
pyiqa>=0.1.9  # Image quality assessment
torchmetrics>=1.2.0

# Utilities
tqdm>=4.66.0
rich>=13.7.0  # Beautiful terminal output
python-dotenv>=1.0.0
pyyaml>=6.0.1
